SLRM: Stabilize the number of workers
Instead of using the number of CPUs we were allocated on the node,
SLURM_CPUS_ON_NODE, we need to use the requested CPUs per GPU times
the number of GPUs. This is because our request is a minimum number
of CPUs, and we might have been allocated more. But we shouldn't
use more than we expected to have access to, otherwise the experiment
won't be reproducible because augmentation depends on the CPU worker
seed.