DEV: Exclude wandb from being detected as first party
Adding src_paths (e77745f) is insufficient, so we need to declare
known_first_party too. This prevents a local directory being
confused for first party code by isort when it is imported,
in particular wandb.

We also needed to add an instruction to replace this when
initializing a repo from the template.